adjust GPSL5 Tests; document functions; gen_code now is able to generate code of size  2^64-1

+++ src/sampling.jl
@@ -37,7 +37,7 @@
     ```
     """
     function gen_sat_code(samples, f, φ₀, f_s, code)
-        code_indices = floor.(Int16, f ./ f_s .* samples .+ φ₀)
+        code_indices = floor.(Int, f ./ f_s .* samples .+ φ₀)
         code_indices .= 1.+ mod.(code_indices, length(code))
         code[code_indices]
     end
